+++ src/main/java/teamseven/echoeco/config/auth/LoginUserArgumentResolver.java
@@ -8,7 +8,7 @@ import org.springframework.web.bind.support.WebDataBinderFactory;
 import org.springframework.web.context.request.NativeWebRequest;
 import org.springframework.web.method.support.HandlerMethodArgumentResolver;
 import org.springframework.web.method.support.ModelAndViewContainer;
-import teamseven.echoeco.user.User;
+import teamseven.echoeco.user.domain.User;
 
 @RequiredArgsConstructor
 @Component
