--- src/main/java/teamseven/echoeco/login/LoginController.java
@@ -1,32 +1,0 @@
-package teamseven.echoeco.login;
-
-import jakarta.servlet.http.HttpServletRequest;
-import jakarta.servlet.http.HttpSession;
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PostMapping;
-
-@RequiredArgsConstructor
-@Controller
-public class LoginController {
-
-    @GetMapping("/")
-    public String index() {
-        return "login/index";
-    }
-
-    @GetMapping("/user/login")
-    public String login() {
-        return "login/loginPage";
-    }
-
-    @PostMapping("/logout")
-    public String logout(HttpServletRequest request) {
-        HttpSession session = request.getSession(false);
-        if (session != null) {
-            session.removeAttribute("user");
-        }
-        return "redirect:/";
-    }
-}
