+++ src/main/java/teamseven/echoeco/login/controller/LoginController.java
@@ -1,0 +1,20 @@
+package teamseven.echoeco.login.controller;
+
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Controller;
+import org.springframework.web.bind.annotation.GetMapping;
+
+@RequiredArgsConstructor
+@Controller
+public class LoginController {
+
+    @GetMapping("/")
+    public String index() {
+        return "login/index";
+    }
+
+    @GetMapping("/user/login")
+    public String login() {
+        return "login/loginPage";
+    }
+}
