+++ src/main/java/teamseven/echoeco/user/controller/UserController.java
@@ -1,14 +1,17 @@
-package teamseven.echoeco.user;
+package teamseven.echoeco.user.controller;
 
+import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
-import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.ModelAttribute;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
-import teamseven.echoeco.login.CustomOAuth2UserService;
+import teamseven.echoeco.user.service.CustomOAuth2UserService;
+import teamseven.echoeco.user.domain.Dto.UserDto;
+import teamseven.echoeco.user.domain.Role;
+import teamseven.echoeco.user.domain.User;
 
 @Controller
 @RequiredArgsConstructor
@@ -26,17 +29,17 @@ public class UserController {
     @GetMapping("/users/{userId}/edit")
     public String updateUserRoleForm(@PathVariable("userId") Long userId, Model model) {
         User user = customOAuth2UserService.findOneById(userId);
-        UserForm userForm = new UserForm();
-        userForm.updateUserForm(user);
-        model.addAttribute("userForm", userForm);
+        UserDto userDto = new UserDto();
+        userDto.updateUserForm(user);
+        model.addAttribute("userDto", userDto);
         model.addAttribute("roles", Role.getAll());
         return "users/updateUserRole";
     }
 
     @PostMapping("/users/{userId}/edit")
     public String updateUserRole(@PathVariable("userId") Long userId,
-                                 @Validated @ModelAttribute("userForm") UserForm userForm) {
-        customOAuth2UserService.updateUserRole(userId, userForm);
+                                 @Valid @ModelAttribute("userDto") UserDto userDto) {
+        customOAuth2UserService.updateUserRole(userId, userDto);
         return "redirect:/users";
     }
