+++ src/main/java/teamseven/echoeco/user/domain/OAuthAttributes.java
@@ -1,4 +1,4 @@
-package teamseven.echoeco.user;
+package teamseven.echoeco.user.domain;
 
 import java.util.Map;
 import lombok.Builder;
