+++ src/main/resources/templates/layouts/menu.html
@@ -57,7 +57,7 @@
                   </g>
                 </svg>
               </span>
-            <span class="app-brand-text demo menu-text fw-bolder ms-2">ㄷEcho Eco</span>
+            <span class="app-brand-text demo menu-text fw-bolder ms-2">Echo Eco</span>
         </a>
 
         <a href="javascript:void(0);" class="layout-menu-toggle menu-link text-large ms-auto d-block d-xl-none">
